requirements.txt
# Main dependencies
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
pydantic-settings==2.5.2

# Optional dependencies for proxy (SOCKS)
# Uncomment the following line if you use a SOCKS5 proxy
# pysocks==1.7.1
# requests[socks]>=2.32.3  # Alternative for SOCKS support
